+++ core/linux/common.cpp
@@ -130,109 +130,6 @@ void install_fault_handler (void)
 #endif
 }
 
-#if !defined(TARGET_NO_THREADS)
-
-//Thread class
-cThread::cThread(ThreadEntryFP* function,void* prm)
-{
-	Entry=function;
-	param=prm;
-}
-
-void cThread::Start()
-{
-	verify(hThread == NULL);
-	if (pthread_create( (pthread_t*)&hThread, NULL, Entry, param))
-	{
-		die("Thread creation failed");
-	}
-}
-
-void cThread::WaitToEnd()
-{
-	if (hThread != NULL)
-	{
-		pthread_join((pthread_t)hThread,0);
-		hThread = NULL;
-	}
-}
-
-//End thread class
-#endif
-
-//cResetEvent Calss
-cResetEvent::cResetEvent(bool State,bool Auto)
-{
-	//sem_init((sem_t*)hEvent, 0, State?1:0);
-	verify(State==false&&Auto==true);
-	pthread_mutex_init(&mutx, NULL);
-	pthread_cond_init(&cond, NULL);
-}
-cResetEvent::~cResetEvent()
-{
-	//Destroy the event object ?
-
-}
-void cResetEvent::Set()//Signal
-{
-	pthread_mutex_lock( &mutx );
-	state=true;
-    pthread_cond_signal( &cond);
-	pthread_mutex_unlock( &mutx );
-}
-void cResetEvent::Reset()//reset
-{
-	pthread_mutex_lock( &mutx );
-	state=false;
-	pthread_mutex_unlock( &mutx );
-}
-bool cResetEvent::Wait(u32 msec)//Wait for signal , then reset
-{
-	pthread_mutex_lock( &mutx );
-	if (!state)
-	{
-		struct timespec ts;
-#if HOST_OS == OS_DARWIN
-		// OSX doesn't have clock_gettime.
-		clock_serv_t cclock;
-		mach_timespec_t mts;
-
-		host_get_clock_service(mach_host_self(), CALENDAR_CLOCK, &cclock);
-		clock_get_time(cclock, &mts);
-		mach_port_deallocate(mach_task_self(), cclock);
-		ts.tv_sec = mts.tv_sec;
-		ts.tv_nsec = mts.tv_nsec;
-#else
-		clock_gettime(CLOCK_REALTIME, &ts);
-#endif
-		ts.tv_sec += msec / 1000;
-		ts.tv_nsec += (msec % 1000) * 1000000;
-		while (ts.tv_nsec > 1000000000)
-		{
-			ts.tv_nsec -= 1000000000;
-			ts.tv_sec++;
-		}
-		pthread_cond_timedwait( &cond, &mutx, &ts );
-	}
-	bool rc = state;
-	state=false;
-	pthread_mutex_unlock( &mutx );
-
-	return rc;
-}
-void cResetEvent::Wait()//Wait for signal , then reset
-{
-	pthread_mutex_lock( &mutx );
-	if (!state)
-	{
-		pthread_cond_wait( &cond, &mutx );
-	}
-	state=false;
-	pthread_mutex_unlock( &mutx );
-}
-
-//End AutoResetEvent
-
 #include <errno.h>
 
 void VArray2::LockRegion(u32 offset,u32 size)
